add different color fish
